feat : 포인트 결제 충전 기능 개발

+++ user/src/main/java/com/zerobase/user/point/domain/model/PointEntity.java
@@ -19,5 +19,5 @@ public class PointEntity extends BaseEntity {
 
     private Long userId;
 
-    private Integer point;
+    private Long point;
 }
